+++ example/main.dart
@@ -92,7 +92,7 @@ class User {
         isAutoIncrement: true,
         isPrimaryKey: true,
       ),
-      'name': TypeString(),
+      'name': TypeVarchar(),
       'date': TypeDateTime(),
     },
   );